fixed stuck seal claim issue, reverting util class changes
