+++ src/main/java/uk/gov/hmcts/reform/civil/handler/callback/camunda/docmosis/GenerateClaimFormForSpecCallbackHandler.java
@@ -16,11 +16,11 @@ import uk.gov.hmcts.reform.civil.model.documents.DocumentMetaData;
 import uk.gov.hmcts.reform.civil.service.DeadlinesCalculator;
 import uk.gov.hmcts.reform.civil.service.Time;
 import uk.gov.hmcts.reform.civil.service.docmosis.sealedclaim.SealedClaimFormGeneratorForSpec;
-import uk.gov.hmcts.reform.civil.service.documentmanagement.DocumentUtil;
 import uk.gov.hmcts.reform.civil.service.stitching.CivilDocumentStitchingService;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
@@ -44,7 +44,6 @@ public class GenerateClaimFormForSpecCallbackHandler extends CallbackHandler {
     private final DeadlinesCalculator deadlinesCalculator;
     private final CivilDocumentStitchingService civilDocumentStitchingService;
     private final FeatureToggleService toggleService;
-    private final DocumentUtil documentUtil;
 
     @Override
     public String camundaActivityId(CallbackParams callbackParams) {
@@ -78,7 +77,7 @@ public class GenerateClaimFormForSpecCallbackHandler extends CallbackHandler {
             callbackParams.getParams().get(BEARER_TOKEN).toString()
         );
 
-        List<DocumentMetaData> documentMetaDataList = documentUtil.fetchDocumentsFromCaseData(caseData, sealedClaim);
+        List<DocumentMetaData> documentMetaDataList = fetchDocumentsFromCaseData(caseData, sealedClaim);
         if (documentMetaDataList.size() > 1) {
             CaseDocument stitchedDocument = civilDocumentStitchingService.bundle(
                 documentMetaDataList,
@@ -95,5 +94,34 @@ public class GenerateClaimFormForSpecCallbackHandler extends CallbackHandler {
         return AboutToStartOrSubmitCallbackResponse.builder()
             .data(caseDataBuilder.build().toMap(objectMapper))
             .build();
+    }
+
+    public List<DocumentMetaData> fetchDocumentsFromCaseData(CaseData caseData, CaseDocument caseDocument) {
+        List<DocumentMetaData> documentMetaDataList = new ArrayList<>();
+        DocumentMetaData documentMetaData = new DocumentMetaData(
+            caseData.getSpecClaimTemplateDocumentFiles(),
+            "doc1",
+            "doc1"
+        );
+
+        documentMetaDataList.add(new DocumentMetaData(caseDocument.getDocumentLink(),
+                                                      "Sealed Claim form",
+                                                      LocalDate.now().toString()));
+        if (caseData.getSpecClaimTemplateDocumentFiles() != null) {
+            documentMetaDataList.add(new DocumentMetaData(
+                caseData.getSpecClaimTemplateDocumentFiles(),
+                "Claim timeline",
+                LocalDate.now().toString()
+            ));
+        }
+        if (caseData.getSpecClaimDetailsDocumentFiles() != null) {
+            documentMetaDataList.add(new DocumentMetaData(
+                caseData.getSpecClaimDetailsDocumentFiles(),
+                "Supported docs",
+                LocalDate.now().toString()
+            ));
+        }
+
+        return documentMetaDataList;
     }
 }
